+++ src/model_diagnostics/calibration/_compute.py
@@ -11,115 +11,12 @@ from model_diagnostics._utils.array import (
     get_sorted_array_names,
     length_of_first_dimension,
     length_of_second_dimension,
-    validate_2_arrays,
     validate_same_first_dimension,
 )
 from model_diagnostics._utils.binning import bin_feature
 from model_diagnostics._utils.partial_dependence import compute_partial_dependence
 
-
-def identification_function(
-    y_obs: npt.ArrayLike,
-    y_pred: npt.ArrayLike,
-    *,
-    functional: str = "mean",
-    level: float = 0.5,
-) -> np.ndarray:
-    r"""Canonical identification function.
-
-    Identification functions act as generalised residuals. See [Notes](#notes) for
-    further details.
-
-    Parameters
-    ----------
-    y_obs : array-like of shape (n_obs)
-        Observed values of the response variable.
-        For binary classification, y_obs is expected to be in the interval [0, 1].
-    y_pred : array-like of shape (n_obs)
-        Predicted values of the `functional`, e.g. the conditional expectation of
-        the response, `E(Y|X)`.
-    functional : str
-        The functional that is induced by the identification function `V`. Options are:
-
-        - `"mean"`. Argument `level` is neglected.
-        - `"median"`. Argument `level` is neglected.
-        - `"expectile"`
-        - `"quantile"`
-
-    level : float
-        The level of the expectile of quantile. (Often called \(\alpha\).)
-        It must be `0 < level < 1`.
-        `level=0.5` and `functional="expectile"` gives the mean.
-        `level=0.5` and `functional="quantile"` gives the median.
-
-    Returns
-    -------
-    V : ndarray of shape (n_obs)
-        Values of the identification function.
-
-    Notes
-    -----
-    [](){#notes}
-
-    The function \(V(y, z)\) for observation \(y=y_{pred}\) and prediction
-    \(z=y_{pred}\) is a strict identification function for the functional \(T\), or
-    induces the functional \(T\) as:
-
-    \[
-    \mathbb{E}[V(Y, z)] = 0\quad \Leftrightarrow\quad z\in T(F) \quad \forall
-    \text{ distributions } F
-    \in \mathcal{F}
-    \]
-
-    for some class of distributions \(\mathcal{F}\). Implemented examples of the
-    functional \(T\) are mean, median, expectiles and quantiles.
-
-    | functional | strict identification function \(V(y, z)\)           |
-    | ---------- | ---------------------------------------------------- |
-    | mean       | \(z - y\)                                            |
-    | median     | \(\mathbf{1}\{z \ge y\} - \frac{1}{2}\)              |
-    | expectile  | \(2 \mid\mathbf{1}\{z \ge y\} - \alpha\mid (z - y)\) |
-    | quantile   | \(\mathbf{1}\{z \ge y\} - \alpha\)                   |
-
-    For `level` \(\alpha\).
-
-    References
-    ----------
-    `[Gneiting2011]`
-
-    :   T. Gneiting.
-        "Making and Evaluating Point Forecasts". (2011)
-        [doi:10.1198/jasa.2011.r10138](https://doi.org/10.1198/jasa.2011.r10138)
-        [arxiv:0912.0902](https://arxiv.org/abs/0912.0902)
-
-    Examples
-    --------
-    >>> identification_function(y_obs=[0, 0, 1, 1], y_pred=[-1, 1, 1 , 2])
-    array([-1,  1,  0,  1])
-    """
-    y_o: np.ndarray
-    y_p: np.ndarray
-    y_o, y_p = validate_2_arrays(y_obs, y_pred)
-
-    if functional in ("expectile", "quantile") and (level <= 0 or level >= 1):
-        msg = f"Argument level must fulfil 0 < level < 1, got {level}."
-        raise ValueError(msg)
-
-    if functional == "mean":
-        return y_p - y_o
-    elif functional == "median":
-        return np.greater_equal(y_p, y_o) - 0.5
-    elif functional == "expectile":
-        return 2 * np.abs(np.greater_equal(y_p, y_o) - level) * (y_p - y_o)
-    elif functional == "quantile":
-        return np.greater_equal(y_p, y_o) - level
-    else:
-        allowed_functionals = ("mean", "median", "expectile", "quantile")
-        msg = (
-            f"Argument functional must be one of {allowed_functionals}, got "
-            f"{functional}."
-        )
-        raise ValueError(msg)
+from ._identification import identification_function
 
 
 def compute_bias(
