+++ src/model_diagnostics/xai/__init__.py
@@ -1,4 +1,4 @@
-from .permutation_importance import compute_permutation_importance
-from .plots import plot_permutation_importance
+from ._permutation_importance import compute_permutation_importance
+from ._plots import plot_permutation_importance
 
 __all__ = ["compute_permutation_importance", "plot_permutation_importance"]
